homework.py cog
